+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/sessionize/Dependencies.kt
@@ -21,7 +21,7 @@ object Libs {
         const val junitJupiter = "5.5.2"
         const val kotlinCoroutinesRetrofit = "1.1.0"
         const val moshi = "1.15.2"
-        const val okhttp = "3.12.5"
+        const val okhttp = "4.12.0"
         const val retrofit = "2.6.2"
         const val threetenbp = "1.7.1"
     }